Rearrange TH uses, to avoid cross-chunk references
When Template Haskell is used to splice definitions into a module, that module
is no longer compiled 'all at once': instead, the part before any TH is
compiled, the part between the first two splices is compiled, and so on until
the whole module is done. The downside of splitting up the compilation like this
is that definitions in earlier 'chunks' cannot reference definitions from later
'chunks', since they won't be found during compilation. Hence why I've had to
rearrange some definitions in the suite/Functionality.hs file.